+++ game/hud/src/components/UI/TabbedDialog/config.ts
