+++ game/hud/src/components/UI/TabbedDialogWithSideMenu.tsx
