+++ game/hud/src/widgets/Settings/components/Key.tsx
