+++ game/hud/src/widgets/Settings/components/SettingsPanel.tsx
