+++ game/hud/src/widgets/Settings/tabs/Interface.tsx
@@ -5,23 +5,24 @@
  */
 
 import * as React from 'react';
-import { DialogTab, DialogButton } from '../../../widgets/UI/TabbedDialog';
-import { SideMenu, MenuOption } from '../../../widgets/UI/SideMenu';
+import { DialogTab, DialogButton } from '../../../components/UI/TabbedDialog';
+import { SideMenu, MenuOption } from '../../../components/UI/SideMenu';
 import { ComingSoon } from '../panels/ComingSoon';
 import * as BUTTON from './buttons';
+import * as OPTION from './options';
 
-const options: MenuOption[] = [];
+const options: MenuOption[] = [OPTION.SKILLS, OPTION.CHAT];
 const buttons: DialogButton[] = [BUTTON.APPLY, BUTTON.CANCEL];
 
-interface AddonsSettingsProps {
+interface InterfaceSettingsProps {
   onCancel: () => void;
 }
 
 /* tslint:disable:function-name */
-export function AddonSettings(props: AddonsSettingsProps) {
+export function InterfaceSettings(props: InterfaceSettingsProps) {
   return (
     <DialogTab buttons={buttons}>
-      <SideMenu name='settings' id='addons' options={options}>{
+      <SideMenu name='settings' id='interface' options={options}>{
         (option: MenuOption) => {
           return <ComingSoon/>;
         }
@@ -30,4 +31,4 @@ export function AddonSettings(props: AddonsSettingsProps) {
   );
 }
 
-export default AddonSettings;
+export default InterfaceSettings;
