+++ game/hud/src/widgets/Settings/tabs/buttons.ts
@@ -4,7 +4,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-import { DialogButton } from '../../../widgets/UI/TabbedDialog';
+import { DialogButton } from '../../../components/UI/TabbedDialog';
 
 export const APPLY: DialogButton = { label: 'Apply' };
 export const DEFAULT: DialogButton = { label: 'Default' };
