+++ game/hud/src/widgets/Settings/utils/configVars.ts
