+++ game/hud/src/widgets/Settings/utils/keyboard.ts
