+++ game/hud/src/widgets/Settings/utils/samples/graphicsConfig.ts
