+++ game/hud/src/widgets/Settings/utils/samples/inputConfig.ts
