+++ game/hud/src/widgets/Settings/utils/samples/keybindsConfig.ts
