--- game/hud/src/widgets/UI/index.ts
@@ -1,12 +1,0 @@
-export {
-  TabbedDialogWithSideMenu,
-  DialogDefinition,
-  DialogButton,
-  MenuOption,
-  DialogContent,
-} from '../../widgets/UI/TabbedDialogWithSideMenu';
-export { Box } from '../../widgets/UI/Box';
-export { SubHeading } from '../../widgets/UI/SubHeading';
-export { Field } from '../../widgets/UI/Field';
-export { CheckBoxField } from '../../widgets/UI/CheckBoxField';
-export { SliderField } from '../../widgets/UI/SliderField';
